Fix: Move tests back into appropriate directory
