+++ tests/Framework/Constraint/ArraySubsetTest.php
@@ -10,6 +10,8 @@
 
 namespace PHPUnit\Framework\Constraint;
 
+use PHPUnit\Framework\ExpectationFailedException;
+
 class ArraySubsetTest extends ConstraintTestCase
 {
     /**
