+++ tests/Framework/Constraint/LessThanTest.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * This file is part of sebastian/phpunit-framework-constraint.
+ * This file is part of PHPUnit.
  *
  * (c) Sebastian Bergmann <sebastian@phpunit.de>
  *
@@ -10,27 +10,26 @@
 
 namespace PHPUnit\Framework\Constraint;
 
-use PHPUnit\Framework\Constraint;
 use PHPUnit\Framework\ExpectationFailedException;
 use PHPUnit\Framework\TestFailure;
 
-class IsNullTest extends ConstraintTestCase
+class LessThanTest extends ConstraintTestCase
 {
-    public function testConstraintIsNull()
+    public function testConstraintLessThan()
     {
-        $constraint = new IsNull();
+        $constraint = new LessThan(1);
 
-        $this->assertFalse($constraint->evaluate(0, '', true));
-        $this->assertTrue($constraint->evaluate(null, '', true));
-        $this->assertEquals('is null', $constraint->toString());
+        $this->assertTrue($constraint->evaluate(0, '', true));
+        $this->assertFalse($constraint->evaluate(1, '', true));
+        $this->assertEquals('is less than 1', $constraint->toString());
         $this->assertCount(1, $constraint);
 
         try {
-            $constraint->evaluate(0);
+            $constraint->evaluate(1);
         } catch (ExpectationFailedException $e) {
             $this->assertEquals(
                 <<<EOF
-Failed asserting that 0 is null.
+Failed asserting that 1 is less than 1.
 
 EOF
                 ,
@@ -43,17 +42,17 @@ EOF
         $this->fail();
     }
 
-    public function testConstraintIsNull2()
+    public function testConstraintLessThan2()
     {
-        $constraint = new IsNull();
+        $constraint = new LessThan(1);
 
         try {
-            $constraint->evaluate(0, 'custom message');
+            $constraint->evaluate(1, 'custom message');
         } catch (ExpectationFailedException $e) {
             $this->assertEquals(
                 <<<EOF
 custom message
-Failed asserting that 0 is null.
+Failed asserting that 1 is less than 1.
 
 EOF
                 ,
