+++ tests/Framework/Constraint/RegularExpressionTest.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * This file is part of sebastian/phpunit-framework-constraint.
+ * This file is part of PHPUnit.
  *
  * (c) Sebastian Bergmann <sebastian@phpunit.de>
  *
@@ -13,25 +13,23 @@ namespace PHPUnit\Framework\Constraint;
 use PHPUnit\Framework\ExpectationFailedException;
 use PHPUnit\Framework\TestFailure;
 
-class IsEmptyTest extends ConstraintTestCase
+class RegularExpressionTest extends ConstraintTestCase
 {
-    public function testConstraintIsEmpty()
+    public function testConstraintRegularExpression()
     {
-        $constraint = new IsEmpty();
+        $constraint = new RegularExpression('/foo/');
 
-        $this->assertFalse($constraint->evaluate(['foo'], '', true));
-        $this->assertTrue($constraint->evaluate([], '', true));
-        $this->assertFalse($constraint->evaluate(new \ArrayObject(['foo']), '', true));
-        $this->assertTrue($constraint->evaluate(new \ArrayObject([]), '', true));
-        $this->assertEquals('is empty', $constraint->toString());
+        $this->assertFalse($constraint->evaluate('barbazbar', '', true));
+        $this->assertTrue($constraint->evaluate('barfoobar', '', true));
+        $this->assertEquals('matches PCRE pattern "/foo/"', $constraint->toString());
         $this->assertCount(1, $constraint);
 
         try {
-            $constraint->evaluate(['foo']);
+            $constraint->evaluate('barbazbar');
         } catch (ExpectationFailedException $e) {
             $this->assertEquals(
                 <<<EOF
-Failed asserting that an array is empty.
+Failed asserting that 'barbazbar' matches PCRE pattern "/foo/".
 
 EOF
                 ,
@@ -44,16 +42,17 @@ EOF
         $this->fail();
     }
 
-    public function testConstraintIsEmpty2()
+    public function testConstraintRegularExpression2()
     {
-        $constraint = new IsEmpty();
+        $constraint = new RegularExpression('/foo/');
 
         try {
-            $constraint->evaluate(['foo'], 'custom message');
+            $constraint->evaluate('barbazbar', 'custom message');
         } catch (ExpectationFailedException $e) {
             $this->assertEquals(
                 <<<EOF
-custom message\nFailed asserting that an array is empty.
+custom message
+Failed asserting that 'barbazbar' matches PCRE pattern "/foo/".
 
 EOF
                 ,
