+++ tests/Framework/ConstraintTest.php
@@ -10,13 +10,12 @@
 
 namespace PHPUnit\Framework;
 
-use PHPUnit\Framework\Constraint;
 use PHPUnit\Framework\Constraint\Count;
 use PHPUnit\Framework\Constraint\SameSize;
 use PHPUnit\Framework\Constraint\TraversableContains;
 use PHPUnit\Util\Filter;
 
-class ConstraintTest extends ConstraintTestCase
+class ConstraintTest extends TestCase
 {
     public function testConstraintArrayNotHasKey()
     {
